libtest/libtestlib/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anchor-keyexchange"
version = "0.2.0"
description = "lib for automating key exchange"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-can",
  "keyexchange",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
